Add files via upload
Hi, I just worked on the country-wise data, this should work so all we need to do is changing the path of the data which needs to be changed as per your environment. Can you please someone run with your could Jupyter notebook? 

We can discuss what we can do to make it better in the last 1h 30 mins.